fix: Exclude digit from non-string non-number JSON
The previous RegExp could cause super-linear backtracking when the JSON
string contains digits outside of a string which are not part of a valid
number.  Since the JSON grammar does not allow digits outside of strings
and numbers, don't allow matching them.  This avoids the potential for
super-linear backtracking.

Thanks
[`regexp/no-super-linear-backtracking`](https://ota-meshi.github.io/eslint-plugin-regexp/rules/no-super-linear-backtracking.html)!

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>